Replace sprintf with snprintf
This commit replaces sprintf which is unsafe as it can result in an
overflow with snprintf. On MacOS this will generate a deprecated
declaration warning using the latest clang.

As these buffers are static array declarations, sizeof can be used to
calculate the size of the buffer, however if replaced with malloc sizeof
will no longer be correct.

Signed-off-by: Harry Ramsey <harry.ramsey@arm.com>